updated EmptyActivity to expect Task from loadUserProfile; updated PhoneValidationServiceImpl to call createUserProfile when data is available
