+++ app/src/main/java/com/cozo/cozomvp/authentication/validationservice/ValidationData.kt
@@ -1,0 +1,16 @@
+package com.cozo.cozomvp.authentication.validationservice
+
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
+
+open class ValidationData{
+
+    constructor(phoneNumber : String){
+        this.phoneNumber = phoneNumber
+    }
+    constructor(account: GoogleSignInAccount){
+        this.account = account
+    }
+
+    lateinit var phoneNumber : String
+    lateinit var account : GoogleSignInAccount
+}
