+++ app/src/main/java/com/cozo/cozomvp/authentication/validationservice/ValidationService.kt
@@ -1,0 +1,9 @@
+package com.cozo.cozomvp.authentication.validationservice
+
+import com.cozo.cozomvp.authentication.AuthModel
+
+interface ValidationService{
+    fun signUserIn(signInData: ValidationData, authModel: AuthModel)
+    fun linkWithAccount(accountData: ValidationData, authModel: AuthModel)
+    fun signUserOut(authModel: AuthModel)
+}
